Add files via upload
Biased Data
While working on this particular dataset, I came across a situation of Data Biaseness. What is a Biased Data?
--> Basically Statistical bias is a term used to describe statistics that don't provide an accurate representation of the population. Some data is flawed because the sample of people it surveys doesn't accurately represent the population.
For example in this case, the target variable is Binary(0,1); But in the dataset of 58592 entries,
0 --> 54844
1 --> 3748
This shows that data is biased towards zeros

So what if the data is biased?
--> If the data is biased the ML algorithm predictions will be dominated by the majority class objects and consequently the minority class is neglected.
In this dataset the majority class is zero, so no matter which algorithm you use-It is going to predict zero for any record.

But the key thing in such datasets is that the minority class is the most important one from the business perspective.

So how to handle this situation?
--> There are multiple methods to tackle data biaseness Some of them are
1. Undersampling the majority class
2. Oversampling the minority class
3. SMOTE - Synthetic Minority Oversampling Technique
4. Use of ensemble for undersampling

I tried to implement all these methods below.